requirements.txt
flask
loguru
pymysql
flask-migrate
flask-sqlalchemy
flask_jwt_extended
